Refactor marketplace add into shared core flow
