+++ codex-rs/core/src/plugins/mod.rs
@@ -6,6 +6,7 @@ mod installed_marketplaces;
 mod manager;
 mod manifest;
 mod marketplace;
+mod marketplace_add;
 mod mentions;
 mod remote;
 mod render;
@@ -58,6 +59,10 @@ pub use marketplace::MarketplacePluginInstallPolicy;
 pub use marketplace::MarketplacePluginPolicy;
 pub use marketplace::MarketplacePluginSource;
 pub use marketplace::validate_marketplace_root;
+pub use marketplace_add::MarketplaceAddError;
+pub use marketplace_add::MarketplaceAddOutcome;
+pub use marketplace_add::MarketplaceAddRequest;
+pub use marketplace_add::add_marketplace;
 pub use remote::RemotePluginFetchError;
 pub use remote::fetch_remote_featured_plugin_ids;
 pub(crate) use render::render_explicit_plugin_instructions;
